UTs on 64bit OS, Part II
git-svn-id: https://svn.code.sf.net/p/quex/code/trunk@4613 db9deaaf-66c7-40b0-811a-ad7aae109365

+++ quex/code_base/buffer/TESTS/load/TEST/backward.c
@@ -41,7 +41,7 @@ main(int argc, char**argv)
 
     if( argc > 1 && strcmp(argv[1], "--hwut-info") == 0 ) {
         printf("Buffer_load_backward: (BPC=%i, FB=%i);\n",
-               sizeof(QUEX_TYPE_LEXATOM),
+               (int)sizeof(QUEX_TYPE_LEXATOM),
                (int)QUEX_SETTING_BUFFER_MIN_FALLBACK_N);
         printf("CHOICES: ");
         switch( QUEX_SETTING_BUFFER_MIN_FALLBACK_N ) {
